Revert "have the rest of the SDK call on the new functions"
This reverts commit 8ffd20f90cb222e42273f6d411347b4315acea9a.
