+++ swift-sdk/Internal/LocalStorageProtocol.swift
@@ -14,8 +14,8 @@ protocol LocalStorageProtocol {
     var offlineMode: Bool { get set }
     func getAttributionInfo(currentDate: Date) -> IterableAttributionInfo?
     func save(attributionInfo: IterableAttributionInfo?, withExpiration expiration: Date?)
-    func getLastPushPayload(_ currentDate: Date) -> [AnyHashable: Any]?
-    func saveLastPushPayload(_ payload: [AnyHashable: Any]?, withExpiration expiration: Date?)
+    func getPayload(currentDate: Date) -> [AnyHashable: Any]?
+    func save(payload: [AnyHashable: Any]?, withExpiration: Date?)
     func upgrade()
 }
